Formally removing posibilyty to create parent files.
It just doesnt work like I would like it to work.  Future iventigation
required.

If variable set to tru, it should allow to "missing" tasks in given path
for refiling.  The problem is when we are using more than one file.
Then emacs gets confused, to which file write those new tasks (since it
can create any linie of taks, it could be any file).